Split cogclasses into visa and guilded cog
